perf(rust, python): add specialized boolean aggregation for min/max
