+++ polars/polars-arrow/src/kernels/take_agg.rs
@@ -1,5 +1,5 @@
 //! kernels that combine take and aggregations.
-use arrow::array::{PrimitiveArray, Utf8Array};
+use arrow::array::{Array, BooleanArray, PrimitiveArray, Utf8Array};
 use arrow::types::NativeType;
 use num_traits::{NumCast, ToPrimitive};
 
@@ -162,4 +162,100 @@ pub unsafe fn take_agg_utf8_iter_unchecked_no_null<
         .into_iter()
         .map(|idx| arr.value_unchecked(idx))
         .reduce(|acc, str_val| f(acc, str_val))
+}
+
+/// Take kernel for single chunk and an iterator as index.
+/// # Safety
+/// caller must ensure iterators indexes are in bounds
+#[inline]
+pub unsafe fn take_min_bool_iter_unchecked_nulls<I: IntoIterator<Item = usize>>(
+    arr: &BooleanArray,
+    indices: I,
+    len: IdxSize,
+) -> Option<bool> {
+    let mut null_count = 0 as IdxSize;
+    let validity = arr.validity().unwrap();
+
+    for idx in indices {
+        if validity.get_bit_unchecked(idx) {
+            if !arr.value_unchecked(idx) {
+                return Some(false);
+            }
+        } else {
+            null_count += 1;
+        }
+    }
+    if null_count == len {
+        None
+    } else {
+        Some(true)
+    }
+}
+
+/// Take kernel for single chunk and an iterator as index.
+/// # Safety
+/// caller must ensure iterators indexes are in bounds
+#[inline]
+pub unsafe fn take_min_bool_iter_unchecked_no_nulls<I: IntoIterator<Item = usize>>(
+    arr: &BooleanArray,
+    indices: I,
+) -> Option<bool> {
+    if arr.is_empty() {
+        return None;
+    }
+
+    for idx in indices {
+        if !arr.value_unchecked(idx) {
+            return Some(false);
+        }
+    }
+    Some(true)
+}
+
+/// Take kernel for single chunk and an iterator as index.
+/// # Safety
+/// caller must ensure iterators indexes are in bounds
+#[inline]
+pub unsafe fn take_max_bool_iter_unchecked_nulls<I: IntoIterator<Item = usize>>(
+    arr: &BooleanArray,
+    indices: I,
+    len: IdxSize,
+) -> Option<bool> {
+    let mut null_count = 0 as IdxSize;
+    let validity = arr.validity().unwrap();
+
+    for idx in indices {
+        if validity.get_bit_unchecked(idx) {
+            if arr.value_unchecked(idx) {
+                return Some(true);
+            }
+        } else {
+            null_count += 1;
+        }
+    }
+    if null_count == len {
+        None
+    } else {
+        Some(false)
+    }
+}
+
+/// Take kernel for single chunk and an iterator as index.
+/// # Safety
+/// caller must ensure iterators indexes are in bounds
+#[inline]
+pub unsafe fn take_max_bool_iter_unchecked_no_nulls<I: IntoIterator<Item = usize>>(
+    arr: &BooleanArray,
+    indices: I,
+) -> Option<bool> {
+    if arr.is_empty() {
+        return None;
+    }
+
+    for idx in indices {
+        if arr.value_unchecked(idx) {
+            return Some(true);
+        }
+    }
+    Some(false)
 }
